refactored email and password auth

+++ src/utils/api/items-service/index.ts
@@ -1,12 +1,11 @@
 import { ItemsApi } from "../../../api/items";
-import { getApiConfig } from "../../firebase/api-config";
 import { Item } from "../../../api/items";
 
-const itemsApi = new ItemsApi(getApiConfig());
+const itemsApi = new ItemsApi();
 
 /// this will be the way forward for the api calls, the rest of the code should be in the reducers or the components will work on this once carlos reviews the uploads
 export const listItem = async (item: Item) => {
-  const  data  = itemsApi.createItemApiV1Post(item);
+  const  data  = await itemsApi.createItemApiV1Post(item, { headers: { Authorization: `Bearer ${sessionStorage.getItem("firebaseToken")}`} });
   return data;
 }
 
@@ -18,18 +17,18 @@ export const fetchItems = async () => {
 
 export const fetchItem = async (id: string) => {
   // This function fetches a single item from the api
-  const { data } = await itemsApi.getItemByIdApiV1IdGet(id);
+  const { data } = await itemsApi.getItemByIdApiV1IdGet(id, { headers: { Authorization: `Bearer ${sessionStorage.getItem("firebaseToken")}`} });
   return data;
 };
 
 export const updateItemById = async (item: Item, id: string) => {
   // This function updates an item by making a call to the api
-  const { data } = await itemsApi.updateItemApiV1IdPut(item, id);
+  const { data } = await itemsApi.updateItemApiV1IdPut(item, id, { headers: { Authorization: `Bearer ${sessionStorage.getItem("firebaseToken")}`} });
   return data;
 }
 
 export const deleteItemById = async (id: string) => {
   // This function deletes an item by making a call to the api
-  const { data } = await itemsApi.deleteItemApiV1IdDelete(id);
+  const { data } = await itemsApi.deleteItemApiV1IdDelete(id, { headers: { Authorization: `Bearer ${sessionStorage.getItem("firebaseToken")}`} });
   return data;
 }
